+++ src/store/appState/appStateActions.ts
@@ -1,6 +1,7 @@
 import { createAction } from '@reduxjs/toolkit'
 import { Theme } from '../../theme/types'
 import { Locale } from '../../localization/types'
+import { WindowInnerSize } from './useAppState'
 
 export const setBridgeServiceIsInitialized = createAction(
     'appState/setBridgeServiceIsInitialized',
@@ -30,6 +31,28 @@ export const setLocale = createAction(
         return {
             payload: {
                 locale
+            }
+        }
+    }
+)
+
+export const setIsTouchDevice = createAction(
+    'appState/setIsTouchDevice',
+    (isTouchDevice: boolean) => {
+        return {
+            payload: {
+                isTouchDevice: isTouchDevice
+            }
+        }
+    }
+)
+
+export const setWindowInnerSize = createAction(
+    'appState/setWindowInnerSize',
+    (windowInnerSize: WindowInnerSize | null) => {
+        return {
+            payload: {
+                windowInnerSize
             }
         }
     }