+++ src/store/appState/appStateReducer.ts
@@ -1,18 +1,30 @@
 import { createReducer } from '@reduxjs/toolkit'
-import { setBridgeServiceIsInitialized, setTheme, setLocale } from './appStateActions'
+import {
+    setBridgeServiceIsInitialized,
+    setTheme,
+    setLocale,
+    setIsTouchDevice,
+    setWindowInnerSize
+} from './appStateActions'
 import { Theme } from '../../theme/types'
 import { Locale } from '../../localization/types'
 import { getPreferredLocale } from '../../localization/localization'
+import { detectWindowInnerSize } from "../../utilities/utilities";
+import { WindowInnerSize } from './useAppState'
 
 type AppStateReducerState = {
+    isTouchDevice: boolean,
     isUnlocked: boolean,
+    windowInnerSize: WindowInnerSize | null,
     bridgeServiceIsInitialized: boolean,
     theme: Theme,
     locale: Locale
 }
 
 const initialState: AppStateReducerState = {
+    isTouchDevice: false,
     isUnlocked: true,
+    windowInnerSize: detectWindowInnerSize(),
     bridgeServiceIsInitialized: false,
     theme: 'Auto',
     locale: getPreferredLocale()
@@ -31,6 +43,15 @@ const appStateReducer = createReducer(initialState, builder => {
         .addCase(setLocale, (state, action) => {
             state.locale = action.payload.locale
         })
+
+        .addCase(setIsTouchDevice, (state, action) => {
+            state.isTouchDevice = action.payload.isTouchDevice
+        })
+
+        .addCase(setWindowInnerSize, (state, action) => {
+            state.windowInnerSize = action.payload.windowInnerSize
+        })
+
 })
 
 export default appStateReducer